+++ src/core/prelude/date/other.ts
@@ -7,6 +7,7 @@
  */
 
 import extend from 'core/prelude/extend';
+import { deprecate } from 'core/meta/deprecation';
 
 /** @see Date.prototype.clone */
 extend(Date.prototype, 'clone', function (this: Date): Date {
@@ -14,5 +15,6 @@ extend(Date.prototype, 'clone', function (this: Date): Date {
 });
 
 /** @see DateConstructor.getWeekDays */
-extend(Date, 'getWeekDays', () =>
-	[t`Mn`, t`Ts`, t`Wd`, t`Th`, t`Fr`, t`St`, t`Sn`]);
+extend(Date, 'getWeekDays', deprecate(function getWeekDays(): string[] {
+	return [t`Mn`, t`Ts`, t`Wd`, t`Th`, t`Fr`, t`St`, t`Sn`];
+}));