+++ src/core/prelude/number/index.ts
@@ -10,6 +10,8 @@
 // tslint:disable:no-bitwise
 
 import extend from 'core/prelude/extend';
+import { deprecate } from 'core/meta/deprecation';
+import { locale as defaultLocale } from 'core/prelude/i18n';
 
 const
 	second = 1e3,
@@ -160,7 +162,7 @@ const
 
 /** @see NumberConstructor.pad */
 extend(Number.prototype, 'pad', function (
-	this: Number,
+	this: number,
 	targetLength: number = 0,
 	opts?: NumberPadOptions
 ): string {
@@ -189,7 +191,7 @@ extend(Number.prototype, 'round', createRoundingFunction(Math.round));
 extend(Number.prototype, 'ceil', createRoundingFunction(Math.ceil));
 
 function createRoundingFunction(method: Function): Function {
-	return function (this: Number, precision?: number): number {
+	return function (this: number, precision?: number): number {
 		const
 			val = Number(this);
 
@@ -212,34 +214,120 @@ function createRoundingFunction(method: Function): Function {
 //#if runtime has prelude/number/format
 
 const globalOpts = Object.createDict({
+	init: false,
 	decimal: '.',
 	thousands: ','
 });
 
 /** @see NumberConstructor.getOption */
-extend(Number, 'getOption', (key: string) => globalOpts[key]);
+extend(Number, 'getOption', deprecate(function getOption(key: string): string {
+	return globalOpts[key];
+}));
 
 /** @see NumberConstructor.setOption */
-extend(Number, 'setOption', (key: string, value: string) => globalOpts[key] = value);
+extend(Number, 'setOption', deprecate(function setOption(key: string, value: string): void {
+	globalOpts.init = true;
+	globalOpts[key] = value;
+}));
+
+const formatAliases = Object.createDict({
+	'$': 'currency',
+	'$d': 'currencyDisplay',
+	'%': 'percent',
+	'.': 'decimal'
+});
+
+const defaultFormat = Object.createDict(<Intl.NumberFormatOptions>{
+	style: 'decimal',
+	currency: 'USD',
+	currencyDisplay: 'symbol'
+});
+
+const boolAliases = Object.createDict({
+	'+': true,
+	'-': false
+});
+
+const
+	formatCache = Object.createDict<Intl.NumberFormat>();
 
 /** @see Number.prototype.format */
 extend(Number.prototype, 'format', function (
-	this: Number,
-	optsOrLength?: number | NumberFormatOptions
+	this: number,
+	patternOrOpts?: number | string | Intl.NumberFormatOptions,
+	locale: CanArray<string> = defaultLocale.value
 ): string {
-	const
-		opts = <NumberFormatOptions>{...globalOpts};
-
-	let
-		decimalLength;
-
-	if (Object.isObject(optsOrLength)) {
-		Object.assign(opts, optsOrLength);
-		decimalLength = opts.decimalLength;
-
-	} else {
-		decimalLength = optsOrLength;
+	if (patternOrOpts === undefined && !globalOpts.init) {
+		return this.toLocaleString(locale);
 	}
+
+	if (Object.isObject(patternOrOpts)) {
+		return this.toLocaleString(locale, patternOrOpts);
+	}
+
+	if (Object.isString(patternOrOpts)) {
+		const
+			pattern = patternOrOpts,
+			cacheKey = [locale, pattern].join(),
+			cache = formatCache[cacheKey];
+
+		if (cache) {
+			return cache.format(this);
+		}
+
+		const
+			chunks = pattern.split(';'),
+			opts = <Intl.NumberFormatOptions>{};
+
+		for (let i = 0; i < chunks.length; i++) {
+			const
+				el = chunks[i].trim();
+
+			let
+				[key, val] = el.split(':');
+
+			key = key.trim();
+
+			if (val) {
+				val = val.trim();
+			}
+
+			const
+				alias = formatAliases[key];
+
+			let
+				brk = false;
+
+			if (alias) {
+				key = alias;
+
+				if (alias === 'currency') {
+					opts.style = 'currency';
+
+					if (val) {
+						opts.currency = val || defaultFormat.currency;
+					}
+
+					opts.style = 'currency';
+					brk = true;
+				}
+			}
+
+			if (!brk) {
+				if (!val) {
+					val = defaultFormat[key];
+				}
+
+				opts[key] = val in boolAliases ? boolAliases[val] : val;
+			}
+		}
+
+		const formatter = formatCache[cacheKey] = new Intl.NumberFormat(locale, opts);
+		return formatter.format(this);
+	}
+
+	const
+		decimalLength = Number(patternOrOpts);
 
 	const
 		val = Number(this),
@@ -252,7 +340,7 @@ extend(Number.prototype, 'format', function (
 	for (let j = 0, i = int.length; i--;) {
 		if (j === 3) {
 			j = 0;
-			res = opts.thousands + res;
+			res = globalOpts.thousands + res;
 		}
 
 		j++;
@@ -260,7 +348,7 @@ extend(Number.prototype, 'format', function (
 	}
 
 	if (dec?.length) {
-		return res + opts.decimal + dec;
+		return res + globalOpts.decimal + dec;
 	}
 
 	return res;
@@ -277,7 +365,7 @@ function createPostfixGetter(nm: string): PropertyDescriptor {
 }
 
 function createMsFunction(offset: number): Function {
-	const fn = function (this: Number): number {
+	const fn = function (this: number): number {
 		return Number(this) * offset;
 	};